apic: fixing the mwait hang issue during s3 exit
This patch was missed in the pmu driver s3 enabling. This is a very
important patch where we dont disable lapic when we are entering S3.
In cases where there is a mwait failure & scu aborts, pmu interrupt
will not be forwarded to IA as lapic is suspended & system will hang
with bothcpu's in mwait state. hence in this patch we try to have a
seperate lapic_suspend/resume routines specific to our platform.

Original patch was altered by David Cohen in kernel 3.4. This patch
merges the changes done by David and Vishwesh.

Signed-off-by: Vishwesh M Rudramuni <vishwesh.m.rudramuni@intel.com>
Signed-off-by: Sathyanarayanan Kuppuswamy <sathyanarayanan.kuppuswamy@intel.com>