io_apic: Mechanism to signal kernel panic to SCU
IO_apic write generates IPC2 event in SCU. Make a dummy
write to unused location to hit IPC2 event to SCU.
Target of the dummy write is at the end of the RTE table.

SCU can check the value and hit an SCU status dump to SRAM.
This mechanims is always available when kernel is running.
It is not blocked by ipc1 or pmu operations. This is used only
in error situation (panic / warn) to trig SCU to refresh its
status to shared RAM.

Signed-off-by: Samu Onkalo <samu.onkalo@intel.com>
Signed-off-by: Jouni Hogander <jouni.hogander@intel.com>