INTEL-MID: re-structure arch/x86/platform/intel-mid
As Intel rolling out more SoC's after Moorestown, we need to re-structure
the code here in a way that is backward compatible and easy to expand.
This patch implements a flexible way to add SoC files to support new SoC,
also defines board files for different hardware designs that based on a
specific SoC.

This patch does not add any new functional support. It just refactors
the existing code to increase the modularity and decrease the code duplication
for supporting multiple soc's and boards.

The following are main changes,

1. Removed the device specific code from mrst.c and moved them to arch/x86/
   platform/intel-mid/device-libs/platform_<device>.*. A new device file is
   added for every supported device.

2. Moved the board configuration structure (struct devs_id device_ids ) definition
   to board.c. This will provide better scalability for our code to support
   multiple boards.

3. Created a sperate file for medfield specific soc defintions (mfld.c). Any
   code which is not common for all atom soc's needs to be in soc specfific
   file.

4. Moved SFI specific parsing/handling code to intel_mid_sfi.*. This will enable
   us to reuse our intel-mid code for platforms that supports other firmware
   interfaces like ACPI.

Signed-off-by: Sathyanarayanan Kuppuswamy <sathyanarayanan.kuppuswamy@intel.com>