pci: Fix the PCI delays for intel-mid platforms
This basically ensures that d3_delay and d3cold_delay are
not needed to be managed by the kernel as they are taken
care in SCU for intel mid platforms. However, for the platforms
like BYT (non intel-mid) we will go with kernel PCI framework
way of managing these delays without interfering.

Signed-off-by: Catalin Popescu <catalin.popescu@intel.com>
Signed-off-by: Srinidhi Kasagar <srinidhi.kasagar@intel.com>