logging: Add emmc_ipanic driver
On a kernel panic, the emmc_ipanic driver writes the kernel log buffer to an eMMC partition.
The partition is read and copied to /proc/emmc_ipanic_console on next reboot.

The intention is to get a usable call stack in case of kernel BUG(),
when there is no time to flush the dmesg to the serial console or the log journal file.

Modification to the original driver:
- used partition 5 instead of 2
- remove log_buf_clear() which results in a modfication to kernel/printk.c
- revert following commit to solve an init sequence issue between mmc/core and emmc_ipanic:
	"emmc_ipanic: bind to panic partition by iterating block devices"

Signed-off-by: Saadi Maalem <saadi.maalem@intel.com>