watchdog: change KWDG_MSI from 15 to 12
SCU sends MSI 15 when kernel watchdog expires.
Purpose is to trigger a BUG() on kernel side, to get
a call stack before resetting.

The MSI 15 is shared with i2c-6, which is used on EDISON.
Therefore any access to i2c-6 triggers a BUG()

To solve this problem, change the KWDG_MSI to 12.
This MSI is for i2c-3, which is unused on EDISON.

Signed-off-by: Saadi Maalem <saadi.maalem@intel.com>